requirements.txt
Flask==3.1.1
Flask_Bcrypt==1.0.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.6.0
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.60
langchain_openai==0.3.17
langgraph==0.4.5
langsmith==0.2.19
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pymongo==4.13.0
schedule==1.2.0
python-dotenv==1.1.0
Requests==2.32.3
selenium==4.32.0
undetected_chromedriver==3.5.5
sentence-transformers
faiss-cpu
ir-measures
openai
